Add ComboBoxItem and ListItem components for enhanced selection UI
